fix: 회원 생성 시 매핑 및 중복 팀원 선택 처리
