+++ src/main/java/com/nexters/moyeomoyeo/team_building/service/UserService.java
@@ -6,6 +6,7 @@ import com.nexters.moyeomoyeo.common.constant.ExceptionInfo;
 import com.nexters.moyeomoyeo.notification.service.NotificationService;
 import com.nexters.moyeomoyeo.team_building.controller.dto.request.UserRequest;
 import com.nexters.moyeomoyeo.team_building.controller.dto.response.UserInfo;
+import com.nexters.moyeomoyeo.team_building.domain.entity.TeamBuilding;
 import com.nexters.moyeomoyeo.team_building.domain.entity.User;
 import com.nexters.moyeomoyeo.team_building.domain.entity.UserChoice;
 import com.nexters.moyeomoyeo.team_building.domain.repository.UserRepository;
@@ -19,14 +20,16 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class UserService {
 
+	private final TeamBuildingService teamBuildingService;
 	private final NotificationService notificationService;
 	private final UserRepository userRepository;
 
-	private static User makeUser(UserRequest request) {
+	public static User makeUser(Long teamBuildingId, UserRequest request) {
 		return User.builder()
 			.name(request.getName())
 			.position(request.getPosition())
 			.profileLink(request.getProfileLink())
+			.teamBuildingId(teamBuildingId)
 			.build();
 	}
 
@@ -36,7 +39,8 @@ public class UserService {
 
 	@Transactional
 	public UserInfo createUser(String teamBuildingUuid, UserRequest request) {
-		final User user = makeUser(request);
+		final TeamBuilding teamBuilding = teamBuildingService.findByUuid(teamBuildingUuid);
+		final User user = makeUser(teamBuilding.getId(), request);
 
 		final List<UserChoice> choices = createUserChoices(request.getChoices());
 		for (final UserChoice choice : choices) {
@@ -49,7 +53,7 @@ public class UserService {
 		return userInfo;
 	}
 
-	private List<UserChoice> createUserChoices(List<String> teamUuids) {
+	public List<UserChoice> createUserChoices(List<String> teamUuids) {
 		final List<UserChoice> choices = new ArrayList<>();
 		int choiceOrder = 1;
 
@@ -70,7 +74,6 @@ public class UserService {
 
 		userRepository.delete(targetUser);
 		notificationService.broadCast(teamBuildingUuid, "delete-user", userUuid);
-
 
 	}
